refactor: shortestSections 관심사를 PathFactory로 변경
